--- src/cli/prepare-tool/common.ts
@@ -1,1 +1,0 @@
-export const PREPARE_TOOL_TOKEN = Symbol('PREPARE_TOOL_TOKEN');
