+++ src/cli/prepare-tool/services/prepare-tool.service.ts
@@ -1,8 +1,8 @@
-import { Injectable } from '@nestjs/common';
 import { execa } from 'execa';
+import { injectable } from 'inversify';
 import { logger } from '../../utils';
 
-@Injectable()
+@injectable()
 export class PrepareToolService {
   async prepareTool(tools: string[]): Promise<void> {
     logger.info(`Preparing legacy tools ${tools.join(', ')} ...`);
